+++ crates/apimock/src/cmd/mod.rs
@@ -1,1 +1,2 @@
 pub mod match_test;
+pub mod validate;